Add basic sizing options for rendering
Added an explicit method for users to specify an initial size for
rendered string, if one has an oversized query and can do estimations
can pre-allocate the right size or more, I tried adding all kinds of
magic logics to determine this but they ended up making the code quite
convoluted and hard to follow (if done efficiently) so for now, the
user knows best
